+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.6.0] - 2021-04-30
+
 ### Added
 
 - Added resource already exists error.
@@ -188,7 +190,8 @@ validation errors.
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/helmclient/compare/v4.5.0...HEAD
+[Unreleased]: https://github.com/giantswarm/helmclient/compare/v4.6.0...HEAD
+[4.6.0]: https://github.com/giantswarm/helmclient/compare/v4.5.0...v4.6.0
 [4.5.0]: https://github.com/giantswarm/helmclient/compare/v4.4.0...v4.5.0
 [4.4.0]: https://github.com/giantswarm/helmclient/compare/v4.3.0...v4.4.0
 [4.3.0]: https://github.com/giantswarm/helmclient/compare/v4.2.1...v4.3.0